LayoutTests/editing/spelling/spelling-insert-html-expected.txt

The spellchecker shouldn't mark substrings of words after pasting.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS markedText is "zz"
PASS successfullyParsed is true

TEST COMPLETE
foo zz
food.
